+++ src/bin/backend/response_error.rs
@@ -33,16 +33,19 @@ pub enum ResponseError {
 
     /// Unsupported content type requested by client.
     #[error("Unsupported content type: {0}")]
-    #[expect(dead_code)]
     UnsupportedContentType(String),
 
     /// Unauthorized access.
     #[error("Unauthorized")]
-    Unauthorized,
+    NotAuthenticated,
 
     /// Template error.
     #[error("Template error")]
     Tera(#[from] tera::Error),
+
+    /// User's role has insufficient permissions.
+    #[error("Action not permitted")]
+    NotPermitted,
 }
 
 impl IntoResponse for ResponseError {
@@ -58,7 +61,8 @@ impl IntoResponse for ResponseError {
             ResponseError::NotFound(_) => StatusCode::NOT_FOUND,
             ResponseError::UnsupportedContentType(_) => StatusCode::UNSUPPORTED_MEDIA_TYPE,
             ResponseError::InvalidInput(_) => StatusCode::BAD_REQUEST,
-            ResponseError::Unauthorized => StatusCode::UNAUTHORIZED,
+            ResponseError::NotAuthenticated => StatusCode::UNAUTHORIZED,
+            ResponseError::NotPermitted => StatusCode::FORBIDDEN,
         };
         // Send only the opaque description using the display trait, to avoid leaking
         // information